Create dice parser structure
Initial commit of structure and some early foundational elements of the
dice parser. Right now we have two concepts of "grammar" the models that
represent the dice, expressions, etc. and then the actual lexical
grammar for analysis and parsing.

The names of these objects are bound to change and will be better
represented in the future, but the first goal should be to get a mostly
working prototype of the dice parser.