Remove value and improve __init__ in Token class
The __init__ function for the base token class now will automatically
assign values to any kwarg that is passed into a token's constructor,
assuming that the parent class's (Token's) init function is still
invoked. In addition, we've removed the default "value" property from
Tokens, the only token which really cares about the original, raw text,
token is the expression.