Modify operator functions on iterables
This is necessary to account for change in how we are evaluating dice.
When we evaluate a dice term now, the Dice object is what is returned.
We still need to iterate over the Dice term and reconcile the top N or
bottom N values, but we should do this over the dice token's inner
"results" structure.