Remove parsers
These were superfluous--pyparsing is handling our parsing for us so we
don't need an explicit parser. If we find ourselves wanting an explicit
parser we can reintroduce these but with less custom regex logic in
place. We should ensure that we're mostly wrapping the behavior of
pyparsing on the grammar we've defined.