Introduce a "Roll" object
In order to provide more utility to the dice rolling package, we want
to introduce a "Roll" object which can not only capture the result of a
roll but also the individual pieces of the roll so you can represent
the rolls however makes sense for your application.

This is the first stage in the refactor. The Roll object currently
produces the same results as the roll() function, to maintain
consistent functionality. We will add improvements iteratively.